style: 패키지 분리
presentaion과 ui 패키지의 기능이 겹쳐 ui패키지 삭제및 presentaion으로 클래스이동
