+++ app/src/main/java/com/example/ssg_tube/presentaion/home/CategoryVideoAdapter.kt
@@ -1,21 +1,24 @@
-package com.example.ssg_tube.presentaion.ui.home
+package com.example.ssg_tube.presentaion.home
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import com.example.ssg_tube.databinding.RvChannelItemBinding
+import com.example.ssg_tube.databinding.RvCategoryVideoItemBinding
 import com.example.ssg_tube.presentaion.model.DetailModel
 
-class ChannelAdapter(private val items: List<DetailModel>) :
-    RecyclerView.Adapter<ChannelAdapter.ViewHolder>() {
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ChannelAdapter.ViewHolder {
+class CategoryVideoAdapter(private val items: List<DetailModel>) :
+    RecyclerView.Adapter<CategoryVideoAdapter.ViewHolder>() {
+    override fun onCreateViewHolder(
+        parent: ViewGroup,
+        viewType: Int
+    ): ViewHolder {
         val binding =
-            RvChannelItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+            RvCategoryVideoItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
         return ViewHolder(binding)
     }
 
-    override fun onBindViewHolder(holder: ChannelAdapter.ViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         holder.bind(items[position])
     }
 
@@ -23,14 +26,16 @@ class ChannelAdapter(private val items: List<DetailModel>) :
         return items.size
     }
 
-    class ViewHolder(private val binding: RvChannelItemBinding) :
+    class ViewHolder(private val binding: RvCategoryVideoItemBinding) :
         RecyclerView.ViewHolder(binding.root) {
         fun bind(item: DetailModel) {
             binding.apply {
                 Glide.with(ivArea.context)
                     .load(item.thumbnail)
                     .into(ivArea)
+                tvTitle.text = item.title
             }
+
         }
     }
 }
