+++ app/src/main/java/com/example/ssg_tube/presentaion/home/ChannelAdapter.kt
@@ -1,0 +1,36 @@
+package com.example.ssg_tube.presentaion.home
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.bumptech.glide.Glide
+import com.example.ssg_tube.databinding.RvChannelItemBinding
+import com.example.ssg_tube.presentaion.model.DetailModel
+
+class ChannelAdapter(private val items: List<DetailModel>) :
+    RecyclerView.Adapter<ChannelAdapter.ViewHolder>() {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+        val binding =
+            RvChannelItemBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        return ViewHolder(binding)
+    }
+
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+        holder.bind(items[position])
+    }
+
+    override fun getItemCount(): Int {
+        return items.size
+    }
+
+    class ViewHolder(private val binding: RvChannelItemBinding) :
+        RecyclerView.ViewHolder(binding.root) {
+        fun bind(item: DetailModel) {
+            binding.apply {
+                Glide.with(ivArea.context)
+                    .load(item.thumbnail)
+                    .into(ivArea)
+            }
+        }
+    }
+}
