+++ app/src/main/java/com/example/ssg_tube/presentaion/mypage/MyPageViewModel.kt
@@ -1,0 +1,4 @@
+package com.example.ssg_tube.presentaion.mypage
+
+class MyPageViewModel {
+}
