--- app/src/main/java/com/example/ssg_tube/presentaion/ui/mypage/MyPageViewModel.kt
@@ -1,4 +1,0 @@
-package com.example.ssg_tube.presentaion.ui.mypage
-
-class MyPageViewModel {
-}
