+++ app/src/main/res/layout/fragment_home.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/bg_color"
     android:orientation="vertical"
-    tools:context=".presentaion.ui.home.HomeFragment">
+    tools:context=".presentaion.home.HomeFragment">
 
     <ScrollView
         android:layout_width="match_parent"
